packages/backend/requirements.txt
pandas
scikit-learn
lightgbm
joblib
torch
psutil
GPUtil
fastapi
uvicorn
huggingface_hub
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
alembic
python-dotenv
structlog
requests
